calculate duration based on video duration reduce accumulated error by calculating segment duration in doubles around which was up to 1ms per segment

+++ streaming/src/main/java/org/mp4parser/streaming/output/mp4/WriterOutputCallback.java
@@ -3,5 +3,5 @@ package org.mp4parser.streaming.output.mp4;
 import org.mp4parser.streaming.StreamingTrack;
 
 public interface WriterOutputCallback {
-    void onSegmentReady(StreamingTrack track, long duration, boolean isInit, boolean isFooter);
+    void onSegmentReady(StreamingTrack track, double duration, boolean isInit, boolean isFooter);
 }
